Lots more changes to fix warnings generated by -Wall option using gcc compiler on apollo.
Note the src/eos/primitive-solver/logs.hpp file still generates warnings about
strict aliasing, but jacob assures me this is OK and will be fixed in the
future (suggested fix is move to C++20.

+++ src/eos/primitive-solver/eos_hybrid.cpp
@@ -125,7 +125,7 @@ void EOSHybrid<LogPolicy>::ReadTableFromFile(std::string fname) {
     m_min_h = std::numeric_limits<Real>::max();
     // New form of bound based on properties of NQT functions and their
     // departure from 'true' log behaviour
-    for (int in = 0; in < m_nn-1; ++in) {
+    for (size_t in = 0; in < m_nn-1; ++in) {
       Real const nb = exp2_(host_log_nb(in));
       Real log2_e_in   = host_table(ECLOGE,in);
       Real log2_e_inp1 = host_table(ECLOGE,in+1);
